requirements.txt
rich
hydra-core
matplotlib
torch
tqdm
pandas
